Initial commit. Work in progress.
Doesn't work well because API updates rate limit
reset time for too long and in the same time counts
all the requests, that is even the ones which are
already over the 60 seconds limit and therefore
daily limit exceeds too fast.
TODO: Add throttler in order to solve the issue.